requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
sqlalchemy
psycopg2-binary
pydantic[email]
alembic
bcrypt>=4.0.1
motor
pymongo
openai
requests
tenacity